major refactor

+++ src/user.ts
@@ -1,29 +1,30 @@
 import { BaseDocument, K2DB } from "@frogfish/k2db/db";
 import { K2Error, ServiceError } from "@frogfish/k2error";
 import { Ticket } from "./types";
-import debugLib from "debug";
 import { Secure } from "./secure";
 import { K2CreateResponse, K2DeleteResponse } from "./k2types";
+import debugLib from "debug";
+import { Auth } from "./auth";
 const debug = debugLib("k2:rbac:user");
 
-export interface UserDocument extends BaseDocument {
-}
+export interface UserDocument extends BaseDocument {}
 
 export class User {
   constructor(private db: K2DB, private ticket: Ticket) {}
 
-  public async createUser(
-    username: string,
+  public async createUserByEmailAndPassword(
+    email: string,
     password: string
   ): Promise<K2CreateResponse> {
-    await this.validateUsername(username);
+    Auth.allow(this.ticket, ["system"], "v00rhnpvfq3c0iftypxs");
+    await this.validateEmail(email);
     await this.validatePassword(password);
     const passwordHash = await Secure.bcrypt(password);
 
     try {
       return await this.db.create("_users", this.ticket.account, {
-        username,
-        passwordHash,
+        email,
+        password: passwordHash,
       });
     } catch (error) {
       throw new K2Error(
@@ -36,41 +37,101 @@ export class User {
   }
 
   public async get(id: string): Promise<UserDocument> {
-    return (await this.db.get("_users", id)) as UserDocument;
+    try {
+      Auth.allow(this.ticket, ["system", "admin"], "uqg1tpsptlm2c4jwj0f7");
+      return (await this.db.get("_users", id)) as UserDocument;
+    } catch (error) {
+      const err: K2Error = error as K2Error;
+      if (err.error === ServiceError.FORBIDDEN) {
+        // cater for "self" access
+        if (this.ticket.user === id) {
+          return (await this.db.get("_users", id)) as UserDocument;
+        } else {
+          throw error;
+        }
+      } else {
+        throw error;
+      }
+    }
+  }
+
+  public async find(
+    query: any,
+    skip?: number,
+    limit?: number
+  ): Promise<UserDocument[]> {
+    try {
+      Auth.allow(this.ticket, ["system", "admin"], "vk0govi50srii8t5g57j");
+      return (await this.db.find(
+        "_users",
+        query,
+        skip,
+        limit
+      )) as UserDocument[];
+    } catch (error) {
+      const err: K2Error = error as K2Error;
+      if (err.error === ServiceError.FORBIDDEN) {
+        // cater for "self" access
+        query._uuid = this.ticket.user;
+        return (await this.db.find(
+          "_users",
+          query,
+          skip,
+          limit
+        )) as UserDocument[];
+      } else {
+        throw error;
+      }
+    }
   }
 
   public async delete(id: string): Promise<K2DeleteResponse> {
-    return await this.db.delete("_users", id);
+    try {
+      Auth.allow(this.ticket, ["system", "admin"], "s32xod35nda11la5ool5");
+      return await this.db.delete("_users", id);
+    } catch (error) {
+      const err: K2Error = error as K2Error;
+      if (err.error === ServiceError.FORBIDDEN) {
+        // cater for "self" access
+        if (this.ticket.user === id) {
+          return await this.db.delete("_users", id);
+        } else {
+          throw error;
+        }
+      } else {
+        throw error;
+      }
+    }
   }
 
-  public async authenticate(
-    username: string,
+  public async authenticateByEmailAndPassword(
+    email: string,
     password: string
-  ): Promise<UserDocument> {
+  ): Promise<{ id: string }> {
     const user = (await this.db.findOne("_users", {
-      username,
+      email,
     })) as UserDocument;
 
     if (!user) {
       throw new K2Error(
         ServiceError.NOT_FOUND,
-        `User ${username} not found`,
+        `User ${email} not found`,
         "kae0644rh475d12gacct",
         undefined
       );
     }
 
-    const isValid = await Secure.bcryptVerify(password, user.passwordHash);
+    const isValid = await Secure.bcryptVerify(password, user.password);
     if (!isValid) {
       throw new K2Error(
         ServiceError.AUTH_ERROR,
-        "Invalid username or password",
+        "Invalid email or password",
         "04eyjdnj7mwfno5v6xz0",
         undefined
       );
     }
 
-    return user;
+    return { id: user._uuid };
   }
 
   public async generateOTP(username: string): Promise<string> {
@@ -100,67 +161,105 @@ export class User {
     return otp;
   }
 
-  public async updatePassword(
-    username: string,
-    newPassword: string,
-    { oldPassword, otp }: { oldPassword?: string; otp?: string }
-  ): Promise<void> {
-    const user = (await this.db.findOne("_users", {
-      username,
-    })) as UserDocument;
+  public async update(
+    id: string,
+    data: Partial<UserDocument>
+  ): Promise<{ updated: number }> {
+    try {
+      Auth.allow(this.ticket, ["system", "admin"], "bo64xz0ki7s8r1u85ro4");
 
-    if (!user) {
-      throw new K2Error(
-        ServiceError.NOT_FOUND,
-        `User ${username} not found`,
-        "41f2akd901jsiwkq0kxm",
-        undefined
-      );
-    }
-
-    let isAuthenticated = false;
-
-    if (oldPassword) {
-      isAuthenticated = await Secure.bcryptVerify(
-        oldPassword,
-        user.passwordHash
-      );
-    } else if (otp) {
-      if (user.otp !== otp) {
-        isAuthenticated = false;
+      // TODO: Add validation and all the data logic such as password hashing
+      // OTP as well. move complexity from handlers to data layer
+      return await this.db.update("_users", id, data);
+    } catch (error) {
+      // allow for self update
+      const err: K2Error = error as K2Error;
+      if (err.error === ServiceError.FORBIDDEN) {
+        if (this.ticket.user === id) {
+          return await this.db.update("_users", id, data);
+        } else {
+          throw error;
+        }
       } else {
-        isAuthenticated = await this.validateOTP(otp, user.otpTimestamp);
+        throw error;
       }
     }
 
-    if (!isAuthenticated) {
-      throw new K2Error(
-        ServiceError.AUTH_ERROR,
-        "Invalid old password or OTP",
-        "f82m7k9x92kql09ahvzz",
-        undefined
-      );
-    }
+    // public async updatePasswordByEmail(
+    //   email: string,
+    //   newPassword: string,
+    //   { oldPassword, otp }: { oldPassword?: string; otp?: string }
+    // ): Promise<void> {
+    //   const user = (await this.db.findOne("_users", {
+    //     email,
+    //   })) as UserDocument;
 
-    await this.validatePassword(newPassword);
+    //   if (!user) {
+    //     throw new K2Error(
+    //       ServiceError.NOT_FOUND,
+    //       `User ${email} not found`,
+    //       "41f2akd901jsiwkq0kxm",
+    //       undefined
+    //     );
+    //   }
 
-    const newPasswordHash = await Secure.bcrypt(newPassword);
+    //   let isAuthenticated = false;
 
-    await this.db.update("_users", user._uuid, {
-      passwordHash: newPasswordHash,
-      otp: null, // Clear OTP after use
-      otpTimestamp: null,
-    });
+    //   if (oldPassword) {
+    //     isAuthenticated = await Secure.bcryptVerify(
+    //       oldPassword,
+    //       user.passwordHash
+    //     );
+    //   } else if (otp) {
+    //     if (user.otp !== otp) {
+    //       isAuthenticated = false;
+    //     } else {
+    //       isAuthenticated = await this.validateOTP(otp, user.otpTimestamp);
+    //     }
+    //   }
+
+    //   if (!isAuthenticated) {
+    //     throw new K2Error(
+    //       ServiceError.AUTH_ERROR,
+    //       "Invalid old password or OTP",
+    //       "f82m7k9x92kql09ahvzz",
+    //       undefined
+    //     );
+    //   }
+
+    //   await this.validatePassword(newPassword);
+
+    //   const newPasswordHash = await Secure.bcrypt(newPassword);
+
+    //   await this.db.update("_users", user._uuid, {
+    //     passwordHash: newPasswordHash,
+    //     otp: null, // Clear OTP after use
+    //     otpTimestamp: null,
+    //   });
+  }
+
+  public async count(): Promise<{ count: number }> {
+    Auth.allow(this.ticket, ["system", "admin"], "908lywg5p6bz0s2j24j5");
+    return this.db.count("_users", {});
   }
 
   /*********** helpers *************/
 
-  private async validateUsername(username: string) {
-    const existingUser = await this.db.findOne("_users", { username });
+  private async validateEmail(email: string) {
+    if (!email) {
+      throw new K2Error(
+        ServiceError.SERVICE_ERROR,
+        "Email is required",
+        "jsn7m263148b721db0xg",
+        undefined
+      );
+    }
+
+    const existingUser = await this.db.findOne("_users", { email });
     if (existingUser) {
       throw new K2Error(
         ServiceError.SERVICE_ERROR,
-        "Username is already taken",
+        "Email is already taken",
         "jsn7m36314jb721db0xg",
         undefined
       );
@@ -168,6 +267,15 @@ export class User {
   }
 
   private async validatePassword(password: string) {
+    if (!password) {
+      throw new K2Error(
+        ServiceError.SERVICE_ERROR,
+        "Password is required",
+        "aaf7m263148b721db0xg",
+        undefined
+      );
+    }
+
     const minLength = 8;
     const hasUpperCase = /[A-Z]/.test(password);
     const hasLowerCase = /[a-z]/.test(password);
